+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## main
+- Added error code to on office exception
 
 ## v0.9.1
 - Added the api claim back to the request factory